Move updateLayerStyles to layerStyles.ts
